refactor: sessions related components
