+++ src/components/main-content.tsx
@@ -2,7 +2,7 @@ import { observer } from 'mobx-react-lite';
 
 import { Lamp } from '@/components/Lamp/lamp';
 import { Login } from '@/components/login';
-import { Sessions } from '@/components/sessions';
+import { Sessions } from '@/components/Sessions/sessions';
 import { Loader } from '@/components/ui/loader';
 import { lamp } from '@/store/lamp';
 import { user } from '@/store/user';
